requirements.txt
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.2.0
cattrs==24.1.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pyflakes==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.35
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.31.0
virtualenv==20.26.6
wget==3.2
